move database to data layer
move and refactor
